Title: X 달리기, Time: 7ms, Memory: 0MB, Status: Passed - CodeTree

+++ 250129/X 달리기/x-run.cpp
@@ -6,7 +6,7 @@ int main() {
     // Please write your code here.
     int X;
     cin >> X;
-    cout << sqrt(X-1) + sqrt(X-1);
+    cout << (int)(sqrt(X-1) + sqrt(X-1));
     return 0;
 }
